adds search by one date and search between dates
